refactor: kratos error handling
