+++ core/api/src/app/errors.ts
@@ -23,9 +23,9 @@ import * as UserErrors from "@/domain/users/errors"
 import * as WalletInvoiceErrors from "@/domain/wallet-invoices/errors"
 import * as SupportError from "@/domain/support/errors"
 import * as OathkeeperError from "@/domain/oathkeeper/errors"
+import * as KratosErrors from "@/domain/kratos/errors"
 
 import * as LedgerFacadeErrors from "@/services/ledger/domain/errors"
-import * as KratosErrors from "@/services/kratos/errors"
 import * as BriaEventErrors from "@/services/bria/errors"
 import * as SvixErrors from "@/services/svix/errors"
 
@@ -55,8 +55,8 @@ export const ApplicationErrors = {
   ...WalletInvoiceErrors,
   ...SupportError,
   ...OathkeeperError,
-
   ...KratosErrors,
+
   ...LedgerFacadeErrors,
   ...BriaEventErrors,
   ...SvixErrors,
