+++ core/api/src/domain/kratos/index.ts
@@ -1,0 +1,1 @@
+export * from "./errors"
