+++ core/api/src/domain/kratos/index.types.d.ts
@@ -1,0 +1,1 @@
+type KratosError = import("./errors").KratosError
